fix problem with resaving last message after remove this one fix dto add method to TelegramMessageRepository

+++ service-api/src/main/java/greencity/dto/telegram/DeleteTelegramMessageRequest.java
@@ -6,5 +6,5 @@ import lombok.Builder;
 @Builder
 public record DeleteTelegramMessageRequest(
         @NotNull Long chatId,
-        Integer messageId,
+        Long messageId,
         Long assetId) { }
