+++ service-api/src/main/java/greencity/dto/telegram/EditTelegramMessageRequest.java
@@ -8,6 +8,6 @@ import lombok.Builder;
 @Builder
 public record EditTelegramMessageRequest(
         @NotNull Long chatId,
-        @NotNull Integer messageId,
+        @NotNull Long messageId,
         @NotBlank @Size(max = 1000) String newText
 ) { }
